Pulled glabs link code out into shared module. Fixed style bug caused by new insert images method.
